Working 0.0.0.1

+++ src/ecs/handle_gen.rs
@@ -1,5 +1,5 @@
 
 pub(crate) trait HandleGen{
-    fn gen(&mut self) -> u32;
+    fn gen(&mut self) -> u64;
 }
